delete images from server

+++ DogFinder/backend/src/modules/post/post.service.ts
@@ -5,12 +5,22 @@ import { Post } from './post.entity';
 import { CreatePostDto } from './dto/post.create-dto';
 import { UpdatePostDto } from './dto/post.update-dto';
 import { User } from '../user/user.entity';
-
+import { EventEmitter2 } from '@nestjs/event-emitter';
+import {
+  BadRequestException,
+  ForbiddenException,
+  Injectable,
+  InternalServerErrorException,
+  Logger,
+  NotFoundException,
+} from '@nestjs/common';
+import { PostDeletedEvent } from './post.events';
 @Injectable()
 export class PostService {
   constructor(
     @InjectRepository(Post)
-    private readonly postRepository: Repository<Post>
+    private readonly postRepository: Repository<Post>,
+    private readonly eventEmitter: EventEmitter2
   ) {}
 
   async findAll(): Promise<Post[]> {
@@ -35,6 +45,24 @@ export class PostService {
   }
 
   async remove(id: number): Promise<void> {
-    await this.postRepository.delete(id);
+    let post = await this.postRepository.findOne({
+      where: [{ id: id }],
+      relations: ['images'],
+    });
+    if (!post) {
+      throw new BadRequestException('Objava ne postoji');
+    }
+    try {
+      await this.postRepository.delete(id);
+
+      this.eventEmitter.emit(
+        'post.deleted',
+        new PostDeletedEvent(id, post.images)
+      );
+    } catch (error) {
+      Logger.error('Error deleting auction:', error);
+      throw new InternalServerErrorException('Greska u toku brisanja');
+    }
+
   }
 }
